requirements.txt
ijson
urllib3
pyyaml